# backend/shortVideoApiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /comment:
        post:
            tags:
                - CommentService
            description: 创建评论
            operationId: CommentService_CreateComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.CreateCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.CreateCommentResponse'
        delete:
            tags:
                - CommentService
            description: 删除评论
            operationId: CommentService_RemoveComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RemoveCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.RemoveCommentResponse'
    /comment/video:
        get:
            tags:
                - CommentService
            description: 列出视频的评论
            operationId: CommentService_ListComment4Video
            parameters:
                - name: videoId
                  in: query
                  description: '@gotags: json:"videoId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                properties:
                                    code:
                                        type: integer
                                        description: Status code. Zero means success.
                                    msg:
                                        type: string
                                        description: Status message. Could be displayed to user.
                                    data:
                                        $ref: '#/components/schemas/svapi.ListComment4VideoResponse'
...
components:
    schemas:
...
        svapi.Comment:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                user:
                    $ref: '#/components/schemas/svapi.CommentUser'
                replyUser:
                    $ref: '#/components/schemas/svapi.CommentUser'
                content:
                    type: string
                date:
                    type: string
                likeCount:
                    type: string
                replyCount:
                    type: string
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
        svapi.CommentUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        svapi.CreateCommentRequest:
            type: object
            properties:
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                content:
                    type: string
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                replyUserId:
                    type: string
                    description: '@gotags: json:"replyUserId,omitempty,string"'
        svapi.CreateCommentResponse:
            type: object
            properties:
                comment:
                    $ref: '#/components/schemas/svapi.Comment'
...
        svapi.ListComment4VideoResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        svapi.RemoveCommentRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        svapi.RemoveCommentResponse:
            type: object
            properties: {}
...